Ensure references to models with self-referential foreign keys work
While models with self-referential foreign keys worked previously,
models which had foreign keys to them did not. This fixes the check
for a self-referential foreign key such that it doesn't pollute the
pks cache while exporting itself, thus allowing the exports of models
with foreign keys to such self-referential models.