Update linting tools now we can
This appears to be needed for compatibility with newer Python versions.
It's unclear why this wasn't previously showing up in CI as fresh
runs without code changes do now hit issues which this change appears
to resolve.

As we're updating `black` here this commit also includes a layout
change to placate the formatter.