Cope with Django 4 lazily creating 'django_migrations' table
Previously this table was always created, even when no migrations
were populated into it. However since devdata relies on creating
databases without actually running migrations this broke the import
logic and the export tests.

Update the export tests so that they call this when needed and the
import tests so that they expose (and thus validate) this issue too.